Completed basic functionality
Created some very very basic UI for the tasks and the background of the website, then programmed the functionality for adding and removing tasks. Tasks were initially removed by their name, but that was changed to their id, which was then determined by the current date, using the Date.now() function.