Offcanvas, Modals, and Project functionality
The first of a couple of commits today. Added an offcanvas from react-bootstrap that displays all the projects you currently have. Then imported fontawesome favicons so I could use them in the project. Afterwards, I imported a modal from react-bootstrap and configured it so you can type a project name and it will add it to the projects displayed on the offcanvas.