Slight Modifications to Style + Projects
So, I rewrote the dropdown component so that I could use it for more things. Likewise, I used useContext and was able to avoid the issue of prop drilling that I had before. Using useContext in of itself was a pain but it wasn't too bad once I figured out how to add multiple states. I completed the functionality for adding projects to a workspace and made some slight work on adding tasks (based on the project they're in of course). Will finish more of this tomorrow.