Finished (almost) commit
This time I finished rewriting some of the code that I did in the last commit, and then worked on fixing the bugs that arose because I rewrote the code. Afterward, I implemented dates, a easily-legible date format (instead of 2025-01-22T22:37 I wrote January 22nd, 2025 at 10:37 pm). Then I made sure to include the ability to automatically add tasks to a project by having the project selected. Likewise, I changed the page so that when the project is selected in the offcanvas, the page updates to show all of the tasks specific to that project. I also added a button to return to the home page, where you can see all of the tasks, irrespective of their project. Lastly, I added a lil tag to each task when they are in the homepage, to let the  user know what project that specific task belongs to. I also added a calendar button that does nothing as of now. And for some reason the visibility modifier on my css isn't working, so there is a white dot on project pages, where the tag for the project would have been. I will fix this in the next commmit.